make function hana specific
